backend/requirements.txt
django==3.2.13
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-model-utils==4.0.0
kubernetes==11.0.0
gunicorn==20.0.4 
confluent-kafka==1.8.2
channels==3.0.4
daphne==3.0.2
py-solc-x==2.0.2
web3==5.28.0